[OG-92] Add equipment endpoints and chest open

+++ polymorphia-backend/src/main/java/com/agh/polymorphia_backend/dto/response/reward/item/ItemResponseDtoBase.java
@@ -11,7 +11,7 @@ public class ItemResponseDtoBase extends BaseRewardResponseDto {
     private String bonusText;
 
     @NotNull
-    private Long limit;
+    private Integer limit;
 
     @NotNull
     private Boolean isLimitReached;
